+++ packages/svelte/src/config.js
@@ -1,7 +1,0 @@
-let config;
-export function setConfig(c) {
-	config = c;
-}
-export function getConfig() {
-	return config;
-}