+++ packages/svelte/src/index.js
@@ -1,3 +1,2 @@
 import Link from './Link.html';
-import { getConfig, setConfig } from './config';
-export { Link, getConfig, setConfig };
+export { Link };