+++ packages/svelte/CHANGELOG.md
@@ -1,3 +1,7 @@
+## Next
+
+* This package now relies on the user using `svelte/store` to make the Curi config available to components (`new Store({ curi: { config }})`)
+
 ## 1.0.0-alpha.5
 
 * `<Link>` uses `history.navigate`, not `history.update`.